fix input for embedding

+++ GePrompt.py
@@ -21,7 +21,7 @@ class promptGe(nn.Module):
 
         task_embed = torch.Tensor([task_id])
         m = task_embed.expand(x_embed.shape[0], -1)
-        n = self.task_embed_layer(m)
+        n = self.task_embed_layer(m).long()
 
         x_task_embed = torch.cat((x_embed, n), dim=1)
         a = self.generation_layer_1(x_task_embed)
